+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## Unreleased
+- Does not store requests on an instance of `Mitm` any longer.
+- Adds `socket` event to `Mitm`.
+- Updated to work with Node v0.11.12.
+
 ## 0.1.337 (Mar 11, 2014)
 - First private release.